Finalise object model, new config format, modular
This is a fairly chunky commit, which is 95% close to
finalising the Bergamot object model.  This covers:
Hosts, Services, Traps, Clusters and Resources.

The Nagios configuration format is dropped, in favour
of our own configuration format, this was the only sane
route forward.  There is a converter which is capable
of converting a Nagios configuration to Bergamot.

The Bergamot project is more modular and is split up into
several Maven modules.

A new SMS notification engine is added, using Twilio, to
notify by text message.