Create site setup site config and DB import cmd.
Add a DB import command to import configuration into the database.

Update the create site command to setup the site configuration
directory with a template site configuration and to import the
site configutation.  As such create-site now gives a good starting
point with sane default configuration.  With the default configuration
loaded a user will be able to login and start setting up checks.