Support agent routed checks
Support stateful routing where a check needs to be executed on a
specific agent.

In the situation where an agent (probably running on a host) is
connecting into the monitoring system, check executions need to
be routed in a stateful manner.  The agent might connect to a
pool of workers, however checks need to be dispatched to a
specific worker.

Agent routing addresses this by making a workers queue bindings
stateful to the connection of an agent.