Start to push permissions processing into database
We need to be able to process contact permissions in the database,
so that we can compute group and location rollups only for the
checks that contacts can see.

Permissions get denormalised form the recursive tree they exist
into and computed into a flat table which can then be joined against
to check what permissions a given contact has on a object

As such when a team / contact is altered, the permissions will need
to be recalculated.